Timer for MSP430 configured. Folder from ESP32: 009 working for wireless ECG recording and file from MSP430: 011_INTAN_with_timer_send_ECG working for wireless ECG and stimulation caused by button pressed

+++ 0 new common-port/common/INTAN_config.h
@@ -40,9 +40,9 @@ typedef struct{
     /* Parameters that may be needed but not necessarily*/
     uint8_t number_of_stimulations;
     uint32_t resting_time;
-    uint16_t stimulation_time;
-    float stimulation_on_time;
-    float stimulation_off_time;
+    uint32_t stimulation_time;
+    uint32_t stimulation_on_time;
+    uint32_t stimulation_off_time;
